Added the checkbox feature
Added the todo checkbox to showcase if the current todo is complete or not.

This was done by creating a label tag in Todo.js and then bringing the {todo.name} tag within it and then creating the input tag. The type is a checkbox. The 'checked' part is linked to the object 'complete' that was created in App.js by typing out checked={todo.complete}.